W10/requirements.txt
Django==3.1.2
django-debug-toolbar==3.1.1
django-extensions==3.0.9
django-redis==4.12.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
Jinja2==2.11.2
PyJWT==1.7.1
redis==3.5.3
requests==2.24.0
python-dotenv==0.15.0
celery==5.0.5
psycopg2-binary==2.8.6
